feat: create service to call backend less functions
